refactor: connect_cache_pool extraction, posture→output_format rename, BatchContext constructor, dead pad_2d_i64 + meta fragment dedup
