+++ src/lib.rs
@@ -81,8 +81,8 @@ pub mod index;
 pub mod kind;
 pub mod language;
 pub mod note;
+pub mod output_format;
 pub mod parser;
-pub mod posture;
 pub mod reference;
 pub mod splade;
 pub mod store;
